requirements.txt
ultralytic
streamlit
numpy
os
opencv-python